refactoring and add new endpoint to get recommender account details
